# docs/swagger.yaml
basePath: /
definitions:
...
  ark_internal_http_dto.RideAssistantBooking:
    properties:
      order_id:
        type: string
      status:
        type: string
    type: object
  ark_internal_http_dto.RideAssistantResponse:
    properties:
      booking:
        $ref: '#/definitions/ark_internal_http_dto.RideAssistantBooking'
      reply:
        type: string
      session:
        $ref: '#/definitions/ark_internal_http_dto.RideAssistantSession'
      status:
        type: string
    type: object
  ark_internal_http_dto.RideAssistantSession:
    properties:
      id:
        type: string
      known_fields:
        additionalProperties:
          type: string
        type: object
      missing_fields:
        items:
          type: string
        type: array
      stage:
        type: string
    type: object
...
  ark_internal_http_dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      user_id:
        type: string
      user_type:
        type: string
    type: object
  ark_internal_httpx.ErrorBody:
    properties:
      error:
        type: string
    type: object
...
  internal_http_handlers.createUserReq:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      user_type:
        type: string
    required:
    - email
    - name
    - user_type
    type: object
...
  internal_http_handlers.rideAssistantReq:
    properties:
      context_info:
        type: string
      message:
        type: string
      session_id:
        type: string
    required:
    - message
    type: object
...
  internal_http_handlers.updateUserNameReq:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    name: ZooArk Team
  description: 'ZooArk 叫車平台後端 API。除了 POST /api/users、GET /health 和 GET /readyz 之外，所有
    endpoint 都需要 Firebase Auth ID Token（`Authorization: Bearer <token>`）。金額欄位以 TWD
    分（1/100 元）為單位。'
  title: ZooArk API
  version: 0.1.0
paths:
...
  /api/assistant/ride/messages:
    post:
      consumes:
      - application/json
      description: Sends a user message to the ride-assistant state machine. The assistant
        responds with a reply, optionally a session snapshot for follow-up turns,
        and optionally a booking block once a ride has been committed.
      parameters:
      - description: Assistant message payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.rideAssistantReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.RideAssistantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Ride-assistant turn
      tags:
      - AI
...
  /api/me:
    delete:
      description: Hard-deletes the authenticated user's profile. There is no soft-delete
        today, so the row is removed immediately.
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Delete current user
      tags:
      - Auth & User
    get:
      description: Returns the profile of the user identified by the Firebase ID Token
        in the Authorization header.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ark_internal_http_dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Get current user
      tags:
      - Auth & User
    patch:
      consumes:
      - application/json
      description: Patches the display name of the currently authenticated user. Only
        the name field is editable today.
      parameters:
      - description: Updated name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.updateUserNameReq'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Update current user name
      tags:
      - Auth & User
...
  /api/users:
    post:
      consumes:
      - application/json
      description: Creates a user row after Firebase signup. This endpoint is public
        (no auth required) — the server trusts the client-supplied email/phone until
        a follow-up PATCH.
      parameters:
      - description: New user payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.createUserReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ark_internal_http_dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      summary: Create user profile
      tags:
      - Auth & User
...
schemes:
- http
- https
securityDefinitions:
  FirebaseAuth:
    description: Firebase ID Token, formatted as `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"